pyproject.toml
[tool.poetry]
name = "fight-me-backend"
version = "0.1.0"
description = ""
authors = ["Tom Vaughan <tomvaughan77@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
python-socketio = "^5.8.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
fastapi = "^0.95.2"
fastapi-socketio = "^0.0.10"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.2.2"

[tool.black]
line-length = 120
target-version = ['py311']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
